Api/requirements.txt
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
Flask-Migrate==2.5.3
psycopg2-binary==2.8.4
cerberus==1.3.2
Flask-JWT-Extended==3.24.1
bcrypt==3.2.0
Flask-Bcrypt==0.7.1
gunicorn
Pillow
python-resize-image
lorem==0.1.1
pytest==5.3.4
coverage==5.0.3